// pom/equipment/interfaces/equipmentTableRowDataForARMG.ts
export interface equipmentTableRowDataForARMG {
    name: string;
    length: string;
    width: string;
    maxWeight: string;
    liftCapability: string;
    maxTwinWeightDifference: string;
    maxTierHeight: string;
    softwareVersion: string;
    hostName: string;
    automationPort: string;
    stackProfilingPort: string;
    craneId: string;
    boundary: string;
}