// pom/equipment/interfaces/equipmentTableRowDataForQC.ts
export interface equipmentTableRowDataForQC {
    name: string;
    maxWeight: string;
    // liftCapability: string;
    maxTwinWeightDifference: string;
    // portalTrolley: string;
    // platform: string;
    availableLocationsOnPlatform: string;
    // sharedLegSpace: string;
    // berth: string;
    // lane: string;
    typologyIndex: string;
    softwareVersion: string;
    provider: string;
    url: string;
    // boundary: string;
}